:seedling: Convert pinned dependencies to probe
